+++ app/src/main/java/de/passwordvault/viewmodel/activities/DuplicatePasswordEntriesViewModel.java
@@ -3,6 +3,8 @@ package de.passwordvault.viewmodel.activities;
 import androidx.lifecycle.ViewModel;
 import java.util.ArrayList;
 import de.passwordvault.model.analysis.passwords.Password;
+import de.passwordvault.model.entry.EntryAbbreviated;
+import de.passwordvault.model.entry.EntryManager;
 
 
 /**
@@ -19,12 +21,26 @@ public class DuplicatePasswordEntriesViewModel extends ViewModel {
      */
     private ArrayList<Password> passwords;
 
+    /**
+     * Attribute stores the entries that are being displayed to the user.
+     */
+    private ArrayList<EntryAbbreviated> entries;
+
+    /**
+     * If the user clicked on an entry to display, a reference to the entry is stored here. In case
+     * the user deletes the entry, this reference can be used to remove the deleted entry from the
+     * list of entries.
+     */
+    private EntryAbbreviated displayedEntry;
+
 
     /**
      * Constructor instantiates a new view model.
      */
     public DuplicatePasswordEntriesViewModel() {
         passwords = null;
+        entries = null;
+        displayedEntry = null;
     }
 
 
@@ -49,6 +65,44 @@ public class DuplicatePasswordEntriesViewModel extends ViewModel {
             throw new NullPointerException();
         }
         this.passwords = passwords;
+        entries = new ArrayList<>();
+        for (Password password : passwords) {
+            EntryAbbreviated abbreviated = EntryManager.getInstance().getAbbreviated(password.getEntryUuid());
+            if (abbreviated != null) {
+                entries.add(abbreviated);
+            }
+        }
+    }
+
+    /**
+     * Method returns the list of entries to display to the user. This is {@code null}, unless
+     * the passwords have been set through {@link #setPasswords(ArrayList)}.
+     *
+     * @return  List of entries.
+     */
+    public ArrayList<EntryAbbreviated> getEntries() {
+        return entries;
+    }
+
+
+    /**
+     * Method returns the displayed entry.
+     *
+     * @return  Displayed entry.
+     * @see     #displayedEntry
+     */
+    public EntryAbbreviated getDisplayedEntry() {
+        return displayedEntry;
+    }
+
+    /**
+     * Method changes the displayed entry.
+     *
+     * @param displayedEntry    New displayed entry.
+     * @see                     #displayedEntry
+     */
+    public void setDisplayedEntry(EntryAbbreviated displayedEntry) {
+        this.displayedEntry = displayedEntry;
     }
 
 }